requirements.txt
certifi==2018.1.18
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
paho-mqtt==1.3.1
pipenv==11.9.0
psycopg2==2.7.4
pytz==2018.3
virtualenv==15.2.0
virtualenv-clone==0.3.0
